+++ tests/test_strangle_app.py
@@ -4,7 +4,7 @@ from alpaca.data.historical.stock import StockHistoricalDataClient
 from alpaca.data.historical.option import OptionHistoricalDataClient
 from datetime import time
 
-from apps.0dte_app import (
+from apps.zero_dte_app import (
     choose_otm_strangle_contracts,
     submit_strangle,
     monitor_and_exit_strangle,
@@ -39,11 +39,11 @@ def fake_chain():
 def test_choose_otm_strangle(monkeypatch, fake_chain):
     # stub option chain and underlying price
     monkeypatch.setattr(
-        "apps.0dte_app.TradingClient.get_option_contracts",
+        "apps.zero_dte_app.TradingClient.get_option_contracts",
         lambda self, req: fake_chain,
     )
     monkeypatch.setattr(
-        "apps.0dte_app.get_underlying_price",
+        "apps.zero_dte_app.get_underlying_price",
         lambda stock, s: 100.0,
     )
 
@@ -69,7 +69,7 @@ def test_submit_strangle(monkeypatch):
         return r
 
     monkeypatch.setattr(
-        "apps.0dte_app.TradingClient.submit_order", fake_submit
+        "apps.zero_dte_app.TradingClient.submit_order", fake_submit
     )
 
     resp = submit_strangle(trading=None, call_symbol="C1", put_symbol="P1", qty=2)
@@ -91,7 +91,7 @@ def test_monitor_and_exit_strangle(monkeypatch):
         return {"C1": DummyTrade(p), "P1": DummyTrade(p)}
 
     monkeypatch.setattr(
-        "apps.0dte_app.OptionHistoricalDataClient.get_option_latest_trade",
+        "apps.zero_dte_app.OptionHistoricalDataClient.get_option_latest_trade",
         fake_trade,
     )
     calls = []
@@ -107,7 +107,7 @@ def test_monitor_and_exit_strangle(monkeypatch):
         return r
 
     monkeypatch.setattr(
-        "apps.0dte_app.TradingClient.submit_order", fake_submit
+        "apps.zero_dte_app.TradingClient.submit_order", fake_submit
     )
 
     # entry_price_sum=2.0, target_pct=0.5 → target_sum=3.0
